requirements.txt
asgiref==3.6.0
atomicwrites==1.4.1
attrs==22.2.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
cssselect2==0.7.0
defusedxml==0.7.1
Django==2.2.16
django-filter==2.4.0
django-multiselectfield==0.1.12
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
Faker==16.6.1
flake8==5.0.4
fonttools==4.38.0
idna==3.4
importlib-metadata==1.7.0
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
mixer==7.1.2
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pydyf==0.5.0
pyflakes==2.5.0
PyJWT==2.1.0
pyphen==0.13.2
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sorl-thumbnail==12.7.0
sqlparse==0.4.3
tinycss2==1.2.1
toml==0.10.2
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
webencodings==0.5.1
zipp==3.12.0
zopfli==0.2.2
